Convert global features to ts
